+++ internal/entity/transport.go
@@ -11,6 +11,6 @@ type Transport struct {
 	Description   string  `db:"description"`
 	Latitude      float64 `db:"latitude"`
 	Longitude     float64 `db:"longitude"`
-	MinutePrice   float64 `db:"minute_price"`
-	DayPrice      float64 `db:"day_price"`
+	MinutePrice   int64   `db:"minute_price"`
+	DayPrice      int64   `db:"day_price"`
 }
